Add the `PageConfiguration` class
This class is used to define the *Configuration Settings* of a Page,
perform authentication and routing, and improve the ability to retrieve
information about other pages.
This class will replace the existing `pageSettings` module and will
likely be used directly with future changes to routing and views.
- Add the `PageConfiguration` class and supporting `SecurityCondition`
& `ShiftConfiguration` classes.
- Implements the `PageConfigurationFramework` namespace, the classes,
traits, & interfaces that make up the `PageConfiguration`.